3d_model_from_text_or_image/requirements.txt
numpy
matplotlib
pyrender
trimesh
Pillow
torch
transformers
rembg
scikit-image
ipywidgets
shap-e @ git+https://github.com/openai/shap-e.git